Author can edit and see admin's post--Fixed

+++ app/Http/Controllers/Author/PostController.php
@@ -107,8 +107,12 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+        if ($post->user_id != Auth::id()) {
+
+            Toastr::error('Your are not Authorized to this post!!!', 'Error');
+            return redirect()->back();
+        }
         return view('author.post.show',compact('post'));
-        //
     }
 
     /**
@@ -119,7 +123,14 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
-         $categories = Category::all();
+
+        if ($post->user_id != Auth::id()) {
+
+            Toastr::error('Your are not Authorized to this post!!!', 'Error');
+            return redirect()->route('author.post.index');
+        }
+
+        $categories = Category::all();
         $tags = Tag::all();
         return view('author.post.edit', compact('post', 'categories', 'tags'));
     }
@@ -133,6 +144,13 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+
+        if ($post->user_id != Auth::id()) {
+
+            Toastr::error('Your are not Authorized to this post!!!', 'Error');
+            return redirect()->route('author.post.index');
+        }
+
          $this->validate($request,[
             'title'=>'required|unique:posts',
             'image'=>'image|mimes:jpg,jpeg,png,bmp',
@@ -199,6 +217,14 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+
+
+        if ($post->user_id != Auth::id()) {
+
+            Toastr::error('Your are not Authorized to this post!!!', 'Error');
+            return redirect()->route('author.post.index');
+        }
+        
         if (Storage::disk('public')->exists('post/'.$post->image)) {
 
             Storage::disk('public')->delete('post/'.$post->image);
